GigHub.Web/Controllers/Api/AttendancesController.cs
using GigHub.Core;
using GigHub.Core.Dtos;
using Microsoft.AspNet.Identity;
using System.Web.Http;

namespace GigHub.Web.Controllers.Api
{
    [Authorize]
    public class AttendancesController : ApiController
    {
        
        private readonly IUnitOfWork _unitOfWork;

        public AttendancesController(IUnitOfWork unitOfWork )
        {
            _unitOfWork = unitOfWork;
        }

        [HttpPost]
        public IHttpActionResult Attend(AttendanceDto dto)
        {
            string userId = User.Identity.GetUserId();

            if (_unitOfWork.Attendances.GetAttendance(dto.GigId, userId)!=null)
            {
                return BadRequest("The attendance already exists.");
            }

            var attendance = new Attendance
            {
                GigId = dto.GigId,
                AttendeeId = userId
            };
            _unitOfWork.Attendances.Add(attendance);
            _unitOfWork.Complete();

            return Ok();
        }

        [HttpDelete]
        public IHttpActionResult Delete(AttendanceDto dto)
        {
            var userId = User.Identity.GetUserId();
            var attendance = _unitOfWork.Attendances.GetAttendance(dto.GigId, userId);

            if (attendance == null) { 
                return NotFound();
            }

            _unitOfWork.Attendances.Remove(attendance);

            _unitOfWork.Complete();

            return Ok(dto);
        }

    }
}
